+++ tests/analysis/conftest.py
@@ -1,5 +1,6 @@
 from topobank.testing.fixtures import (  # noqa: F401
+    test_analysis_function,
     two_topos,
     two_users,
-    user_three_topographies_three_surfaces_three_tags
+    user_three_topographies_three_surfaces_three_tags,
 )