+++ tests/analysis/test_analyses_removal.py
@@ -2,27 +2,21 @@
 Test whether analyses are recalculated on certain events.
 """
 
-from pathlib import Path
-
 import pytest
-from django.contrib.contenttypes.models import ContentType
 from django.shortcuts import reverse
 
-from topobank.analysis.models import Analysis, AnalysisFunction
+from topobank.analysis.models import Analysis
 from topobank.manager.models import Topography
 from topobank.testing.factories import (
-    FIXTURE_DATA_DIR,
+    SurfaceAnalysisFactory,
     SurfaceFactory,
     Topography1DFactory,
-    UserFactory,
-    SurfaceAnalysisFactory,
     Topography2DFactory,
     TopographyAnalysisFactory,
+    UserFactory,
 )
-from topobank.testing.utils import upload_file
 
 
-@pytest.mark.parametrize("auto_renew", [True, False])
 @pytest.mark.parametrize(
     "changed_values_dict",
     [  # would should be changed in POST request (->str values!)
@@ -53,22 +47,14 @@ from topobank.testing.utils import upload_file
     ],
 )
 @pytest.mark.django_db
-def test_renewal_on_topography_change(
+def test_analysis_removal_on_topography_change(
     api_client,
-    mocker,
-    settings,
     django_capture_on_commit_callbacks,
+    test_analysis_function,
     handle_usage_statistics,
     changed_values_dict,
-    auto_renew,
 ):
     """Check whether methods for renewal are called on significant topography change."""
-    settings.CELERY_TASK_ALWAYS_EAGER = True  # perform tasks locally
-    settings.AUTOMATICALLY_RENEW_ANALYSES = auto_renew
-
-    renew_topo_analyses_mock = mocker.patch(
-        "topobank.analysis.controller.submit_analysis"
-    )
 
     user = UserFactory()
     surface = SurfaceFactory(creator=user)
@@ -80,6 +66,9 @@ def test_renewal_on_topography_change(
         instrument_type=Topography.INSTRUMENT_TYPE_CONTACT_BASED,
         instrument_parameters={"tip_radius": {"value": 1.0, "unit": "mm"}},
     )
+    TopographyAnalysisFactory(subject_topography=topo, function=test_analysis_function)
+
+    assert Analysis.objects.filter(subject_dispatch__topography=topo).count() == 1
 
     api_client.force_login(user)
 
@@ -101,14 +90,11 @@ def test_renewal_on_topography_change(
     }
     changed_data_for_post = initial_data_for_post.copy()
 
-    # Reset mockers
-    # renew_cache_mock.reset_mock()
-    renew_topo_analyses_mock.reset_mock()
-
     # Update data
     changed_data_for_post.update(changed_values_dict)  # here is a change at least
 
-    # if we post the initial data, nothing should have been changed, so no actions should be triggered
+    # If we post the initial data, nothing should have been changed, so no actions
+    # should be triggered
     with django_capture_on_commit_callbacks(execute=True) as callbacks:
         response = api_client.patch(
             reverse("manager:topography-api-detail", kwargs=dict(pk=topo.pk)),
@@ -119,10 +105,11 @@ def test_renewal_on_topography_change(
     assert len(callbacks) == 0
     # Nothing changed, so no callbacks
 
-    renew_topo_analyses_mock.assert_not_called()
+    # Check that analysis still exists
+    assert Analysis.objects.filter(subject_dispatch__topography=topo).count() == 1
 
     #
-    # now we post the changed data, some action (=callbacks) should be triggered
+    # Now we post the changed data, some action (=callbacks) should be triggered
     #
     with django_capture_on_commit_callbacks(execute=True) as callbacks:
         response = api_client.patch(
@@ -132,20 +119,9 @@ def test_renewal_on_topography_change(
     assert response.status_code == 200
 
     assert len(callbacks) == 1
-    # one callbacks on commit expected:
-    #   Renewing topography cache (thumbnail, DZI, etc.)
 
-    if auto_renew:
-        renew_topo_analyses_mock.assert_called()
-        assert renew_topo_analyses_mock.call_count == sum(
-            [
-                x.is_implemented_for_type(ContentType.objects.get_for_model(Topography))
-                for x in AnalysisFunction.objects.all()
-            ]
-        )
-    else:
-        renew_topo_analyses_mock.assert_not_called()
-        assert renew_topo_analyses_mock.call_count == 0  # Never called
+    # Check that the analysis has been removed since it is now deprecated
+    assert Analysis.objects.filter(subject_dispatch__topography=topo).count() == 0
 
 
 @pytest.mark.django_db
@@ -180,66 +156,9 @@ def test_analysis_removal_on_topography_deletion(
 
     # No more topography analyses left
     assert Analysis.objects.filter(subject_dispatch__topography=topo).count() == 0
+
     # No more surface analyses left, because the surface no longer has topographies
-    # The analysis of the surface is not deleting in this test, because the analysis does not actually run.
-    # (Analysis run `on_commit`, but this is never triggered in this test.)
+    # The analysis of the surface is not deleting in this test, because the analysis
+    # does not actually run. (Analysis run `on_commit`, but this is never triggered in
+    # this test.)
     # assert Analysis.objects.filter(subject_dispatch__surface=surface).count() == 0
-
-
-@pytest.mark.parametrize("auto_renew", [True, False])
-@pytest.mark.django_db
-def test_renewal_on_topography_creation(
-    api_client,
-    mocker,
-    settings,
-    handle_usage_statistics,
-    django_capture_on_commit_callbacks,
-    auto_renew,
-):
-    settings.CELERY_TASK_ALWAYS_EAGER = True  # perform tasks locally
-    settings.AUTOMATICALLY_RENEW_ANALYSES = auto_renew
-
-    renew_topo_analyses_mock = mocker.patch(
-        "topobank.analysis.controller.submit_analysis"
-    )
-
-    user = UserFactory()
-    surface = SurfaceFactory(creator=user)
-    api_client.force_login(user)
-
-    #
-    # open first step of wizard: file upload
-    #
-    input_file_path = Path(
-        FIXTURE_DATA_DIR + "/example-2d.npy"
-    )  # maybe use package 'pytest-datafiles' here instead
-    with django_capture_on_commit_callbacks(execute=True) as callbacks:
-        response = upload_file(
-            str(input_file_path),
-            surface.id,
-            api_client,
-            django_capture_on_commit_callbacks,
-            name="topo1",
-            measurement_date="2020-10-21",
-            data_source=0,
-            description="description",
-            size_x=1,
-            size_y=1,
-            unit="nm",
-            height_scale=1,
-            detrend_mode="height",
-        )
-        assert response.data["name"] == "topo1"
-    assert len(callbacks) == 1  # renewing cached quantities
-
-    if auto_renew:
-        renew_topo_analyses_mock.assert_called()
-        assert renew_topo_analyses_mock.call_count == 2 * sum(
-            [
-                x.is_implemented_for_type(ContentType.objects.get_for_model(Topography))
-                for x in AnalysisFunction.objects.all()
-            ]
-        )
-    else:
-        renew_topo_analyses_mock.assert_not_called()
-        assert renew_topo_analyses_mock.call_count == 0