+++ topobank/settings/base.py
@@ -5,9 +5,9 @@ Base settings to build other settings files upon.
 import importlib.metadata
 import random
 import string
-from backports.entry_points_selectable import entry_points
 
 import environ
+from backports.entry_points_selectable import entry_points
 from django.core.exceptions import ImproperlyConfigured
 from watchman import constants as watchman_constants
 
@@ -92,7 +92,7 @@ THIRD_PARTY_APPS = [
     "allauth.socialaccount.providers.orcid",
     "rest_framework",
     "storages",
-    "guardian",
+    "guardian",  # needed for migrations only
     "notifications",
     "tagulous",  # tag-model with hierarchies
     "trackstats",
@@ -102,6 +102,7 @@ THIRD_PARTY_APPS = [
 LOCAL_APPS = [
     # Your stuff: custom apps go here
     "topobank.users.apps.UsersAppConfig",
+    "topobank.authorization.apps.AuthorizationAppConfig",
     "topobank.manager.apps.ManagerAppConfig",
     "topobank.analysis.apps.AnalysisAppConfig",
     "topobank.usage_stats.apps.UsageStatsAppConfig",
@@ -148,7 +149,6 @@ DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
 AUTHENTICATION_BACKENDS = [
     "django.contrib.auth.backends.ModelBackend",
     "allauth.account.auth_backends.AuthenticationBackend",
-    "guardian.backends.ObjectPermissionBackend",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#auth-user-model
 AUTH_USER_MODEL = "users.User"
@@ -208,7 +208,8 @@ print(f"PLUGIN_MIDDLEWARE: {PLUGIN_MIDDLEWARE}")
 MIDDLEWARE += PLUGIN_MIDDLEWARE
 
 MIDDLEWARE += [
-    "topobank.middleware.anonymous_user_middleware",  # we need guardian's kind of anonymous user for API calls
+    # we need guardian's kind of anonymous user for API calls
+    "topobank.middleware.anonymous_user_middleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
@@ -474,14 +475,6 @@ MEDIA_ROOT = ""
 MEDIA_URL = "/media/"
 
 #
-# Settings for django-guardian
-#
-GUARDIAN_MONKEY_PATCH = False
-GUARDIAN_RENDER_403 = True
-# uses template "403.html" by default
-# see https://django-guardian.readthedocs.io/en/stable/configuration.html#guardian-render-403
-
-#
 # Settings for tracking package versions for analyses
 #
 # list of tuples of form (import_name, expression_returning_version_string)
@@ -535,12 +528,6 @@ TRACKED_DEPENDENCIES += [
         "allauth.__version__",
         "BSD 3-Clause",
         "https://django-allauth.readthedocs.io/en/latest/",
-    ),
-    (
-        "guardian",
-        "guardian.__version__",
-        "BSD 3-Clause",
-        "https://django-guardian.readthedocs.io/en/stable/",
     ),
     (
         "storages",
@@ -712,9 +699,3 @@ REQUEST_PROFILER_LOG_TRUNCATION_DAYS = 14
 
 # Upload method
 UPLOAD_METHOD = env("TOPOBANK_UPLOAD_METHOD", default="POST")
-
-# Automatically renew analyses when the topography is upload are changes?
-# (If disabled, analyses will run when requested, i.e. viewed.)
-AUTOMATICALLY_RENEW_ANALYSES = env(
-    "TOPOBANK_AUTOMATICALLY_RENEW_ANALYSES", default=False
-)